Update date format in conditional check
In the conditional check for tasks with no repetition, a different date format is being used compared to the one used during the creation of the task. As a result, the outcome of the expression is always false, causing tasks with no repetition to not be displayed on the homepage.